1.6.10 -Ujednolicenie wysylania komunikatow (WYGRALES/PRZEGRALES), i proby zamykania graczy komputerowych -Po raz n'ty, gdzie n to bardzo duzo, proba naprawienia wyswietlania komunikatow o powodzie przegranej w trybie wojny robotow z widoczna plansza

+++ src/logic/RobotPlayer.java
@@ -173,9 +173,9 @@ public class RobotPlayer {
     }
 
     public void killRobot() {
-        robotProc.destroyForcibly();
-        robotProc.destroy();
-        robotProc = null;
+        if(robotProc.isAlive()) {
+            robotProc.destroyForcibly();
+        }
         reader = null;
         writer = null;
     }
